+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## HEAD
+
+* Make tus-ruby-server fully work with non-rewindable Rack input (@janko-m)
+
 ## 2.1.1 (2018-15-26)
 
 * Rename `:download_url` option to `:redirect_download` (@janko-m)